Only process the first module
The `process` method is called every time a module is defined. In the `red-arrow` gem the `Arrow` module is defined in basically every source file, and it is unnecessary to process the GIR file more than once.

Since the `GObjectIntropsectionHandler` class seems to instantiated again for every module, we need need to store a class variable to know if we processed it already.

Also, this intentially doesn't remember _which_ module we processed. This is because if we generate the documentation for the `red-parquet` gem we don't want to process the GIR file for the `Arrow` module, and the `Parquet` module happens to be defined first.