Improvements for documenting Arrow (#6)
Currently there is no good documentation for the Arrow Ruby bindings.
You need to look at both the Ruby code and the C GLib bindings to try to
piece together which classes and methods are available.

Check out `arrow` and `yard-gobject-introspection` side by side, and run
the following command inside the `arrow/ruby/red-arrow` directory:

```
$ GI_GIR_PATH=/opt/homebrew/share/gir-1.0 yard doc --load ../../../yard-gobject-introspection/lib/yard-gobject-introspection.rb --verbose --debug
```

This generates documentation in the `doc` directory that should contain
all the classes and methods from both the C GLib bindings and the Ruby
code together.

In addition to the `GI_GIR_PATH` environment variable, this adds
information about superclasses, class methods, and subtypes to the
generated documentation. It is still not perfect, but I find it very
useful.

It also makes sure to only process each GIR-file once.